Extract founder allele effects at a single marker from output of qtl2::scan1coef

Extract founder allele effects at a single marker from output of qtl2::scan1coef

get_effects(marker_index, allele_effects_matrix, map, columns = 1:8)

Arguments

marker_index
an integer indicating where in the `map` object the peak position (or position of interest) is located

allele_effects_matrix
output of `qtl2::scan1coef` for a single chromosome

map
a map object for the chromosome of interest

columns
which columns to choose within the `allele_effects_matrix`. Default is 1:8 to reflect 8 founder alleles of Diversity Outbred mice

Value

a vector of 8 founder allele effects at a single marker

a vector of founder allele effects at a single marker

Examples

# set up allele effects matrix
ae <- matrix(dat = rnorm(100 * 8), ncol = 8, nrow = 100)
ae[, 8] <- - rowSums(ae[, 1:7])
colnames(ae) <- LETTERS[1:8]
rownames(ae) <- paste0(1, "_", 1:100)
# set up map
map <- 1:100
names(map) <- rownames(ae)
# call get_effects
get_effects(marker_index = 15, allele_effects_matrix = ae, map = map)

#>           A           B           C           D           E           F 
#>  1.40426569  1.01635979 -0.33658898 -0.58955641  0.05817145 -1.32739903 
#>           G           H 
#>  1.71047207 -1.93572458