Simulate a single multivariate data set consisting of n subjects and d phenotypes for each

Simulate a single multivariate data set consisting of n subjects and d phenotypes for each

sim1(X, B, Sigma)

Arguments

X
design matrix (incorporating genotype probabilities from two loci), dn by df

B
a matrix of allele effects, f rows by d columns

Sigma
dn by dn covariance matrix

Value

a vector of length dn. The first n entries are for trait 1, the second n for trait 2, etc.

Examples

n_mouse <- 20
geno <- rbinom(n = n_mouse, size = 1, prob = 1 / 2)
X <- gemma2::stagger_mats(geno, geno)
B <- matrix(c(1, 2), ncol = 2, nrow = 1)
sim1(X, B, Sigma = diag(2 * n_mouse))

#>               [,1]
#>  [1,] -0.319978820
#>  [2,] -0.278834279
#>  [3,]  1.039317096
#>  [4,]  1.594650797
#>  [5,] -0.979050445
#>  [6,]  0.204207089
#>  [7,]  1.085960972
#>  [8,] -0.349671966
#>  [9,]  1.702971138
#> [10,]  0.728471344
#> [11,]  1.859593181
#> [12,]  3.024155203
#> [13,] -0.337234198
#> [14,]  3.477145067
#> [15,]  2.376264164
#> [16,] -1.473551488
#> [17,] -0.345227990
#> [18,]  0.950180682
#> [19,] -0.146245649
#> [20,]  1.782644366
#> [21,]  0.475705566
#> [22,]  3.756315261
#> [23,] -0.230472754
#> [24,]  1.404155281
#> [25,] -2.231484275
#> [26,]  1.840332640
#> [27,]  0.946685864
#> [28,]  3.223728019
#> [29,]  0.325544925
#> [30,]  0.137240272
#> [31,]  3.454136238
#> [32,]  2.841060978
#> [33,]  0.906640280
#> [34,]  1.311339631
#> [35,]  2.482115591
#> [36,]  1.169676620
#> [37,]  0.344012511
#> [38,]  0.002671474
#> [39,]  3.432658547
#> [40,]  1.645490970