Subset an input object - allele probabilities array or phenotypes matrix or covariates matrix. Kinship has its own subset function

An inputted matrix or 3-dimensional array is first subsetted - by rownames - to remove those subjects who are not in `id2keep`. After that, the object's rows are ordered to match the ordering of subject ids in the vector `id2keep`. This (possibly reordered) object is returned.

subset_input(input, id2keep)

Arguments

input
a matrix of either phenotypes or covariates or array of allele probabilities

id2keep
a character vector of subject ids to identify those subjects that are shared by all inputs

Value

an object resulting from subsetting of `input`. Its rows are ordered per `id2keep`

Examples

# define s_id
s_id <- paste0("s", 1:10)
# set up input matrix
foo <- matrix(data = rnorm(10 * 3), nrow = 10, ncol = 3)
rownames(foo) <- s_id
subset_input(input = foo, id2keep = s_id)

#>            [,1]        [,2]       [,3]
#> s1  -1.87282861  0.83250613  0.2926086
#> s2   0.09132266 -0.91378741  0.3322364
#> s3   0.71236377 -0.27259543 -0.2319348
#> s4  -0.11011949  0.75108655 -0.8158072
#> s5  -1.05157582 -0.07685626  0.2160748
#> s6  -0.89140134  0.14710153 -1.1388658
#> s7  -0.61913804 -0.32123820  0.1401050
#> s8   0.25519905 -1.70314007 -0.6621339
#> s9  -0.43931592 -0.76103844 -0.5343918
#> s10  1.46814921 -0.75902539 -1.4128309