Prepare mytab object for use within scan_pvl R code

Prepare mytab object for use within scan_pvl R code

prep_mytab(d_size, n_snp, pvl = TRUE)

Arguments

d_size
an integer, the number of traits

n_snp
an integer, the number of markers

pvl
logical indicating whether to output dataframe with all d-tuples for a d-QTL scan, or only those models that examine one marker at a time.

Value

a data.frame with d_size + 1 columns and (n_snp)^d_size rows. Last column is NA and named loglik.

Examples

prep_mytab(2, 10)

#>     Var1 Var2 loglik
#> 1      1    1     NA
#> 2      2    1     NA
#> 3      3    1     NA
#> 4      4    1     NA
#> 5      5    1     NA
#> 6      6    1     NA
#> 7      7    1     NA
#> 8      8    1     NA
#> 9      9    1     NA
#> 10    10    1     NA
#> 11     1    2     NA
#> 12     2    2     NA
#> 13     3    2     NA
#> 14     4    2     NA
#> 15     5    2     NA
#> 16     6    2     NA
#> 17     7    2     NA
#> 18     8    2     NA
#> 19     9    2     NA
#> 20    10    2     NA
#> 21     1    3     NA
#> 22     2    3     NA
#> 23     3    3     NA
#> 24     4    3     NA
#> 25     5    3     NA
#> 26     6    3     NA
#> 27     7    3     NA
#> 28     8    3     NA
#> 29     9    3     NA
#> 30    10    3     NA
#> 31     1    4     NA
#> 32     2    4     NA
#> 33     3    4     NA
#> 34     4    4     NA
#> 35     5    4     NA
#> 36     6    4     NA
#> 37     7    4     NA
#> 38     8    4     NA
#> 39     9    4     NA
#> 40    10    4     NA
#> 41     1    5     NA
#> 42     2    5     NA
#> 43     3    5     NA
#> 44     4    5     NA
#> 45     5    5     NA
#> 46     6    5     NA
#> 47     7    5     NA
#> 48     8    5     NA
#> 49     9    5     NA
#> 50    10    5     NA
#> 51     1    6     NA
#> 52     2    6     NA
#> 53     3    6     NA
#> 54     4    6     NA
#> 55     5    6     NA
#> 56     6    6     NA
#> 57     7    6     NA
#> 58     8    6     NA
#> 59     9    6     NA
#> 60    10    6     NA
#> 61     1    7     NA
#> 62     2    7     NA
#> 63     3    7     NA
#> 64     4    7     NA
#> 65     5    7     NA
#> 66     6    7     NA
#> 67     7    7     NA
#> 68     8    7     NA
#> 69     9    7     NA
#> 70    10    7     NA
#> 71     1    8     NA
#> 72     2    8     NA
#> 73     3    8     NA
#> 74     4    8     NA
#> 75     5    8     NA
#> 76     6    8     NA
#> 77     7    8     NA
#> 78     8    8     NA
#> 79     9    8     NA
#> 80    10    8     NA
#> 81     1    9     NA
#> 82     2    9     NA
#> 83     3    9     NA
#> 84     4    9     NA
#> 85     5    9     NA
#> 86     6    9     NA
#> 87     7    9     NA
#> 88     8    9     NA
#> 89     9    9     NA
#> 90    10    9     NA
#> 91     1   10     NA
#> 92     2   10     NA
#> 93     3   10     NA
#> 94     4   10     NA
#> 95     5   10     NA
#> 96     6   10     NA
#> 97     7   10     NA
#> 98     8   10     NA
#> 99     9   10     NA
#> 100   10   10     NA