Perform model fitting for all ordered pairs of markers in a genomic region of interest

Perform model fitting for all ordered pairs of markers in a genomic region of interest

scan_pvl(probs, pheno, kinship, covariates = NULL, start_snp1 = 1, n_snp,
  max_iter = 1e+06, max_prec = 1/1e+08)

Arguments

probs
an array of genotype probabilities for a single chromosome (before eigenrotation)

pheno
a matrix of phenotypes (before eigenrotation)

kinship
a kinship matrix for one chromosome

covariates
a matrix, n subjects by n.cov covariates, where each column is one covariate

start_snp1
index of where to start the scan within probs

n_snp
the number of (consecutive) snps to include in the scan

max_iter
maximum number of iterations for EM algorithm

max_prec
stepwise precision for EM algorithm. EM stops once incremental difference in log likelihood is less than max_prec

Value

a tibble with d + 1 columns. First d columns indicate the genetic data (by listing the marker ids) used in the design matrix; last is log likelihood

Examples

## define probs
probs_pre <- rbinom(n = 100 * 10, size = 1, prob = 1 / 2)
probs <- array(data = probs_pre, dim = c(100, 1, 10))
dimnames(probs)[[3]] <- paste0("Marker", 1:10)
# define Y
Y_pre <- runif(200)
Y <- matrix(data = Y_pre, nrow = 100)
covariates <- matrix(c(runif(99), NA), nrow = 100, ncol = 1)
cov2 <- matrix(c(covariates[1:99], 10), nrow = 100, ncol = 1)
Y2 <- Y
Y2[1, 2] <- NA
scan_pvl(probs = probs, pheno = Y, kinship = diag(100), start_snp1 = 1, n_snp = 10)

#> # A tibble: 100 x 3
#>    Var1     Var2    loglik
#>    <chr>    <chr>    <dbl>
#>  1 Marker1  Marker1  -162.
#>  2 Marker2  Marker1  -173.
#>  3 Marker3  Marker1  -192.
#>  4 Marker4  Marker1  -170.
#>  5 Marker5  Marker1  -198.
#>  6 Marker6  Marker1  -172.
#>  7 Marker7  Marker1  -184.
#>  8 Marker8  Marker1  -177.
#>  9 Marker9  Marker1  -172.
#> 10 Marker10 Marker1  -165.
#> # ... with 90 more rows

scan_pvl(probs = probs, pheno = Y2, kinship = diag(100), start_snp1 = 1, n_snp = 10)

#> 1 subjects dropped due to missing values

#> # A tibble: 100 x 3
#>    Var1     Var2    loglik
#>    <chr>    <chr>    <dbl>
#>  1 Marker1  Marker1  -156.
#>  2 Marker2  Marker1  -170.
#>  3 Marker3  Marker1  -186.
#>  4 Marker4  Marker1  -166.
#>  5 Marker5  Marker1  -191.
#>  6 Marker6  Marker1  -169.
#>  7 Marker7  Marker1  -178.
#>  8 Marker8  Marker1  -174.
#>  9 Marker9  Marker1  -165.
#> 10 Marker10 Marker1  -162.
#> # ... with 90 more rows