Simulate a single multivariate data set consisting of n subjects and d phenotypes for each

Simulate a single multivariate data set consisting of n subjects and d phenotypes for each

sim1(X, B, Sigma)

Arguments

X
design matrix (incorporating genotype probabilities from two loci), dn by df

B
a matrix of allele effects, f rows by d columns

Sigma
dn by dn covariance matrix

Value

a vector of length dn. The first n entries are for trait 1, the second n for trait 2, etc.

Examples

n_mouse <- 20
geno <- rbinom(n = n_mouse, size = 1, prob = 1 / 2)
X <- gemma2::stagger_mats(geno, geno)
B <- matrix(c(1, 2), ncol = 2, nrow = 1)
sim1(X, B, Sigma = diag(2 * n_mouse))

#>              [,1]
#>  [1,]  0.37769313
#>  [2,]  1.92660650
#>  [3,]  1.45435482
#>  [4,]  1.57440044
#>  [5,]  2.05984222
#>  [6,] -0.76422002
#>  [7,]  1.76400192
#>  [8,] -0.30053093
#>  [9,]  0.01175105
#> [10,]  0.85071030
#> [11,] -1.55456583
#> [12,] -0.20399135
#> [13,] -0.19156194
#> [14,]  0.51595739
#> [15,] -0.30128401
#> [16,] -0.82552051
#> [17,]  1.63654057
#> [18,]  1.01253254
#> [19,] -0.90210237
#> [20,]  0.50419825
#> [21,]  0.50782823
#> [22,]  0.24735027
#> [23,] -0.80601772
#> [24,]  3.96075593
#> [25,]  2.18975968
#> [26,] -1.36811281
#> [27,]  2.82786930
#> [28,]  3.12390932
#> [29,]  2.85217358
#> [30,]  2.09160758
#> [31,] -0.70066867
#> [32,] -0.52775293
#> [33,]  0.50559912
#> [34,]  1.76343433
#> [35,]  1.32342217
#> [36,]  1.01357184
#> [37,]  1.42835885
#> [38,] -0.08151085
#> [39,]  1.52673800
#> [40,]  2.60623267