Simulate a single data set consisting of n subjects and 2 phenotypes for each

Simulate a single data set consisting of n subjects and 2 phenotypes for each

sim1(X, B, Vg, Ve, kinship)

Arguments

X
design matrix (incorporating genotype probabilities from two loci), 2n by 2f

B
a matrix of allele effects, f rows by 2 columns

Vg
a genetic covariance matrix, 2 by 2

Ve
an error covariance matrix, 2 by 2

kinship
a chromosome-specific kinship matrix, n by n